job_scheduler/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
croniter==2.0.1
python-dateutil==2.8.2
pytz==2023.3 